+++ src/shared/store/member.store.ts
@@ -1,0 +1,16 @@
+import { Member } from '@/entities/member';
+import { create } from 'zustand';
+
+interface MemberState {
+  member: Member | null;
+  isAuthenticated: boolean;
+  setMember: (member: Member | null) => void;
+  clearMember: () => void;
+}
+
+export const useMemberStore = create<MemberState>((set) => ({
+  member: null,
+  isAuthenticated: false,
+  setMember: (member) => set({ member }),
+  clearMember: () => set({ member: null }),
+}));
