--- src/shared/store/session-store.ts
@@ -1,14 +1,0 @@
-import { Member } from '@/entities';
-import { create } from 'zustand';
-
-interface AuthState {
-  user: Member | null;
-  setUser: (user: Member | null) => void;
-  clearUser: () => void;
-}
-
-export const useAuthStore = create<AuthState>((set) => ({
-  user: null,
-  setUser: (user) => set({ user }),
-  clearUser: () => set({ user: null }),
-}));
